backend_django/requirements.txt
Django==5.1.1
django-ninja==1.3.0
djangorestframework-simplejwt==5.3.1
environs==11.0.0
dj-database-url==2.2.0
email_validator==2.2.0
idna==3.10
dnspython==2.7.0
django-cors-headers==4.4.0
pillow==10.4.0
